+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add support for prepend user property operation.
 * Fix support for 32-bit devices. Switch to using sqlite3.0, and cast return values from sqlite3.
 * Add support for logging events to multiple Amplitude apps. See [Readme](https://github.com/amplitude/Amplitude-iOS#tracking-events-to-multiple-amplitude-apps) for details.
+* Add shorthand method to unset user properties
 
 ## 3.5.0 (January 15, 2016)